Created utils dir. Put zipUtils in there. Calling that in users and places blueprints to handle zips
